+++ tests/Stash/Test/AbstractItemTest.php
@@ -357,6 +357,22 @@ abstract class AbstractItemTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($stash->isHit(), 'isHit returns true for valid data');
     }
 
+    public function testExists()
+    {
+        $stash = $this->testConstruct(array('This', 'Should', 'Fail'));
+        $this->assertFalse($stash->exists(), 'exists returns false for missing data');
+        $data = $stash->get();
+        $this->assertNull($data, 'getData returns null for missing data');
+
+        $key = array('isHit', 'test');
+
+        $stash = $this->testConstruct($key);
+        $stash->set('testString')->save();
+
+        $stash = $this->testConstruct($key);
+        $this->assertTrue($stash->exists(), 'exists returns true for valid data');
+    }
+
     public function testClear()
     {
         // repopulate